Refactor Stack file structure
