+++ packages/experiments/src/components/Stack/examples/Stack.Basic.Example.tsx
@@ -26,8 +26,10 @@ export class StackBasicExample extends React.Component<{}, {}> {
           <Text size="medium">Text 2</Text>
           <Text size="large">Text 3</Text>
         </Stack>
+        <Stack.Item align="start">I am a string</Stack.Item>
         <Stack.Item align="end">
           <Text>I am a stack item!</Text>
+          <Text>I am another stack item</Text>
         </Stack.Item>
       </Stack>
     );
