+++ packages/experiments/src/components/Stack/index.ts
@@ -1,2 +1,2 @@
 export * from './Stack';
-export * from './StackItem';
+export * from './StackItem/StackItem';
